refactor: 우승자를 최종결과값을 갖고있는 FinalResult 클래스에서 반환하도록 수정
